Add the zIndexOffset option to configure the dialog "height"
Previously, the height was computed to be 10 higher than the max height
of all of the input's parents. In some cases where the datepicker needs
to move/scroll around with the page, it may collide with other fixed
elements. In those cases, a configurable z-index is very helpful in
order to keep the datepicker visible.